Calculate Stability Score for a Package

The pkg_stablity function calculates a stability score for a specified package. The score is calculated as the percentage of releases with no breaking changes. A breaking change is defined as either a removed function or removed function parameter. The object also returns the raw data from which the stability score is calculated. This data can be useful for review or even custom analytics.

pkg_stability(pkg, releases = NULL, months = NULL)

Arguments

pkg

The name of the package.

releases

An integer indicating the number of releases to collect stability data for. For example, releases = 10 will return stability data for the last 10 releases of the package. Default is NULL, which means the function will return data for all releases.

months

An integer indicating the number of months back to collect stability data for. For example, months = 24 will collect stability data for the previous 2 years. Default is NULL, meaning there is no limitation on the number of release months, and the function will collect data from all releases.

Value

An stability score object of class "pdiff_score". The object contains the stability score, plus other useful information such as the release and version ranges, the number of releases, and number of breaking releases.

Details

The pkg_stability function is used to get an overall feel for the stability of a package. The stability score and other information returned by the function are based on data retrived from CRAN. The function compares each version of a package and determines if any functions or function parameters have been removed from the previous release. If a release removes functions or parameters contained in the previous release, it is flagged as a "breaking release" or "breaking change".

The stability score is calculated as the percentage of non-breaking releases. For example, if a package has 10 releases, and one breaking release, the stability score will be .9, or 90\

The stability assessment is a categorization of the stability score. The aim of the assessment is to provide a general evaluation of the package, whether it is stable or unstable. The assessment has 5 categories:

Perfect: Stability score of 100. Package has never had a breaking release in its entire history.

Very Stable: Stability score between 95 and 100. Package has had a breaking release, but very rarely.

Stable: Stability score between 90 and 95. The package has had some breaking releases, but it is still rather rare.

Somewhat unstable: Stability score between 80 and 90. The package sometimes has a breaking release, at a rate of about 1 in 5.

Unstable: Stability score below 80. The package breaks more frequently than every 5th release. This frequency of breaking releases is considered unstable.

As can be seen above, the stability assessment is weighted heavily toward the high end. This categorization is intentionally designed to encourage package stability, and discourage the number of breaking changes.

Stability Data

Stability calculations are based on a data frame of stability information gathered from each package release. The data frame will contain one row for each release. Each row will contains comparison information against the prior release. The columns are organized as follows:

Version: The version number of the release.

FileName: The file name of the package.

Release: The release date of the package.

Size: The size of the package file on disk.

AF: The number of added functions.

AP: The number of added parameters.

RF: The number of removed functions.

RP: The number of removed parameters.

BC: Whether the release had any breaking changes. Breaking changes are removed functions or removed parameters. Values are zero or one. The value one means the release had breaking changes.

TF: The total number of functions in the package.

The above data will be included in the function print out. It an also be accessed on the returned object.

Disclaimers

Note that a "breaking release" does not currently factor in the number of functions deprecated. One deprecated function counts the same as 10 deprecated functions within a release. Current methodology also does not differentiate between removed functions and removed parameters. They are both considered "breaking", and counted equally.

Current methodology also does not factor in changed functions. The reason is simply that it is much more difficult to identify whether or not a changed function is breaking. That is to say, you may still have breaking changes that are not identified by the pkgdiff package.

Performance

Performance of the pkg_stability function can vary greatly depending on the package selected. The

See also